Feature/gewichtung

+++ NoVe/Models/Note.cs
@@ -7,7 +7,7 @@ namespace NoVe.Models
         public int UserId { get; set; }
         public int FachId { get; set; }
         public int FachbereichId { get; set; }
-        public float Notenwert { get; set; }
+        public double Notenwert { get; set; }
         public int Semester { get; set; }
         public int StudentAlreadyChanged { get; set; }
     }
